# aspnetcore/src/IO.Swagger/wwwroot/swagger-original.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "API for POOP Group 12 contacts app",
    "version" : "1.0.0",
    "title" : "POOP Group 12 API"
  },
  "host" : "crm-group12.com",
  "basePath" : "/PHP/",
  "schemes" : [ "http" ],
  "paths" : {
    "/AmAlive.php" : {
      "post" : {
        "tags" : [ "Users" ],
        "summary" : "Tests endpoint",
        "description" : "Checks if alive",
        "operationId" : "amAlive",
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "I am alive"
          },
          "404" : {
            "description" : "I am dead :("
          }
        }
      }
    },
    "/Login.php" : {
      "post" : {
        "tags" : [ "Users" ],
        "summary" : "Logs in to contacts app",
        "description" : "Logs in to contacts app",
        "operationId" : "doLogin",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "loginItem",
          "description" : "Login",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Login"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          },
          "404" : {
            "description" : "URL Not Found"
          }
        }
      }
    },
    "/CreateUser.php" : {
      "post" : {
        "tags" : [ "Users" ],
        "summary" : "Adds user to user table",
        "description" : "Adds user into user table",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "newUser",
          "description" : "New user",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/User"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully created a new user"
          },
          "404" : {
            "description" : "Could not create user"
          }
        }
      }
    }
  },
  "definitions" : {
    "Login" : {
      "type" : "object",
      "required" : [ "login", "password" ],
      "properties" : {
        "login" : {
          "type" : "string",
          "example" : "SamH"
        },
        "password" : {
          "type" : "string",
          "example" : "Test"
        }
      },
      "example" : {
        "password" : "Test",
        "login" : "SamH"
      }
    },
    "User" : {
      "type" : "object",
      "required" : [ "email", "firstName", "lastName", "login", "password" ],
      "properties" : {
        "login" : {
          "type" : "string",
          "example" : "testUser"
        },
        "password" : {
          "type" : "string",
          "example" : "testPass"
        },
        "firstName" : {
          "type" : "string",
          "example" : "Mike"
        },
        "lastName" : {
          "type" : "string",
          "example" : "Wazowski"
        },
        "email" : {
          "type" : "string",
          "example" : "mike.waz@gmail.com"
        }
      },
      "example" : {
        "firstName" : "Mike",
        "lastName" : "Wazowski",
        "password" : "testPass",
        "login" : "testUser",
        "email" : "mike.waz@gmail.com"
      }
    }
  }
}